services/ingest/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
xarray==2024.11.0
cfgrib==0.9.14.1
fsspec==2024.10.0
gcsfs==2024.10.0
s3fs==2024.10.0
numpy==2.1.3
google-cloud-bigquery==3.27.0
google-cloud-storage==2.19.0
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
eccodes==2.38.3
